Install msw-native gcc iff required

+++ install_msw.sh
@@ -282,16 +282,17 @@ do
     export PATH="$minimal_path"
 
     # For Cygwin or WSL, install and use this msw-native compiler.
-    # Install it for other build types, too, even if only for
-    # validating the installation procedure.
-    mingw_dir=/opt/lmi/${LMI_COMPILER}_${LMI_TRIPLET}/gcc_msw
-    [ -d "$mingw_dir" ] && rm --force --recursive "$mingw_dir"
-    if   [ "i686-w64-mingw32"   = "$LMI_TRIPLET" ]; then
-      make "$coefficiency" --output-sync=recurse -f install_mingw32.make
-    elif [ "x86_64-w64-mingw32" = "$LMI_TRIPLET" ]; then
-      make "$coefficiency" --output-sync=recurse -f install_mingw.make
-    else
-      printf 'No MinGW compiler for this triplet.\n'
+    if [ "Cygwin" = "$platform" ] || [ "WSL" = "$platform" ]
+    then
+        mingw_dir=/opt/lmi/${LMI_COMPILER}_${LMI_TRIPLET}/gcc_msw
+        [ -d "$mingw_dir" ] && rm --force --recursive "$mingw_dir"
+        if   [ "i686-w64-mingw32"   = "$LMI_TRIPLET" ]; then
+          make "$coefficiency" --output-sync=recurse -f install_mingw32.make
+        elif [ "x86_64-w64-mingw32" = "$LMI_TRIPLET" ]; then
+          make "$coefficiency" --output-sync=recurse -f install_mingw.make
+        else
+          printf 'No MinGW compiler for this triplet.\n'
+        fi
     fi
 
     ./install_xml_libraries.sh
